vector-viz/requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
qdrant-client==1.12.1
numpy==1.26.4
umap-learn==0.5.6
pymongo==4.9.2
